+++ packages/suite-base/src/panels/ThreeDeeRender/Interactions/constants.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2026 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+/** Grace period (ms) before tooltip hides after entities clear. */
+export const HOVER_TOOLTIP_GRACE_PERIOD_MS = 400;
+
+/** Delay (ms) before tooltip hides after mouse leaves the tooltip. */
+export const HOVER_TOOLTIP_LEAVE_DELAY_MS = 300;
+
+/** Pixel offset from cursor to tooltip edge. */
+export const HOVER_TOOLTIP_OFFSET_PX = 6;
