+++ packages/suite-base/src/panels/ThreeDeeRender/RendererOverlay.test.tsx
@@ -111,18 +111,19 @@ describe("<RendererOverlay /> hover wiring", () => {
 
   it("maps hovered selections into HoverTooltip entities and absolute client position", async () => {
     const canvas = document.createElement("canvas");
-    canvas.getBoundingClientRect = jest.fn(() =>
-      ({
-        left: 100,
-        top: 200,
-        right: 500,
-        bottom: 600,
-        width: 400,
-        height: 400,
-        x: 100,
-        y: 200,
-        toJSON: () => "",
-      }) as DOMRect,
+    canvas.getBoundingClientRect = jest.fn(
+      () =>
+        ({
+          left: 100,
+          top: 200,
+          right: 500,
+          bottom: 600,
+          width: 400,
+          height: 400,
+          x: 100,
+          y: 200,
+          toJSON: () => "",
+        }) as DOMRect,
     );
 
     renderOverlay(canvas);
@@ -149,13 +150,15 @@ describe("<RendererOverlay /> hover wiring", () => {
     };
 
     act(() => {
-      cb?.([
-        {
-          renderable,
-          instanceIndex: undefined,
-        },
-      ] as any,
-      { x: 10, y: 20 } as any);
+      cb?.(
+        [
+          {
+            renderable,
+            instanceIndex: undefined,
+          },
+        ] as any,
+        { x: 10, y: 20 } as any,
+      );
     });
 
     await waitFor(() => {
@@ -186,23 +189,26 @@ describe("<RendererOverlay /> hover wiring", () => {
     );
 
     // nested objects are excluded
-    expect(entity.metadata).not.toEqual(expect.arrayContaining([{ key: "nested", value: "[object Object]" }]));
+    expect(entity.metadata).not.toEqual(
+      expect.arrayContaining([{ key: "nested", value: "[object Object]" }]),
+    );
   });
 
   it("uses instanceDetails when instanceIndex is provided", async () => {
     const canvas = document.createElement("canvas");
-    canvas.getBoundingClientRect = jest.fn(() =>
-      ({
-        left: 0,
-        top: 0,
-        right: 100,
-        bottom: 100,
-        width: 100,
-        height: 100,
-        x: 0,
-        y: 0,
-        toJSON: () => "",
-      }) as DOMRect,
+    canvas.getBoundingClientRect = jest.fn(
+      () =>
+        ({
+          left: 0,
+          top: 0,
+          right: 100,
+          bottom: 100,
+          width: 100,
+          height: 100,
+          x: 0,
+          y: 0,
+          toJSON: () => "",
+        }) as DOMRect,
     );
 
     renderOverlay(canvas);
@@ -243,18 +249,19 @@ describe("<RendererOverlay /> hover wiring", () => {
 
   it("clears hovered entities when selections are empty", async () => {
     const canvas = document.createElement("canvas");
-    canvas.getBoundingClientRect = jest.fn(() =>
-      ({
-        left: 0,
-        top: 0,
-        right: 100,
-        bottom: 100,
-        width: 100,
-        height: 100,
-        x: 0,
-        y: 0,
-        toJSON: () => "",
-      }) as DOMRect,
+    canvas.getBoundingClientRect = jest.fn(
+      () =>
+        ({
+          left: 0,
+          top: 0,
+          right: 100,
+          bottom: 100,
+          width: 100,
+          height: 100,
+          x: 0,
+          y: 0,
+          toJSON: () => "",
+        }) as DOMRect,
     );
 
     renderOverlay(canvas);
@@ -267,6 +274,53 @@ describe("<RendererOverlay /> hover wiring", () => {
     });
 
     await waitFor(() => {
+      expect(mockLastHoverTooltipProps.entities).toEqual([]);
+    });
+  });
+
+  it("does not create tooltip entries for selections with no topic and no metadata", async () => {
+    const canvas = document.createElement("canvas");
+    canvas.getBoundingClientRect = jest.fn(
+      () =>
+        ({
+          left: 0,
+          top: 0,
+          right: 100,
+          bottom: 100,
+          width: 100,
+          height: 100,
+          x: 0,
+          y: 0,
+          toJSON: () => "",
+        }) as DOMRect,
+    );
+
+    renderOverlay(canvas);
+
+    const cb = mockRendererEventCallbacks.get("renderableHovered");
+    expect(cb).toBeDefined();
+
+    const renderable = {
+      topic: undefined,
+      name: "af949d5a-8243-4e53-8b39-dfb05aac50ba",
+      details: jest.fn(() => undefined),
+      instanceDetails: jest.fn(() => undefined),
+    };
+
+    act(() => {
+      cb?.(
+        [
+          {
+            renderable,
+            instanceIndex: undefined,
+          },
+        ] as any,
+        { x: 10, y: 20 } as any,
+      );
+    });
+
+    await waitFor(() => {
+      expect(mockLastHoverTooltipProps).toBeDefined();
       expect(mockLastHoverTooltipProps.entities).toEqual([]);
     });
   });
