+++ packages/suite-base/src/panels/ThreeDeeRender/constants.ts
@@ -12,3 +12,6 @@ export const PANEL_STYLE: React.CSSProperties = {
 export const MAX_TRANSFORM_MESSAGES = 10_000;
 
 export const DEFAULT_FOLLOW_MODE = "follow-pose";
+
+// Throttle hover GPU picking to reduce load (10 Hz)
+export const HOVER_PICK_THROTTLE_MS = 100;
